resolve conflict according to PR comments
