+++ Unity/Assets/Scripts/MessageReceiver.cs.meta
@@ -1,11 +1,2 @@
 fileFormatVersion: 2
-guid: cc938a55d0b684c968d5491d412eaffa
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
+guid: 4e5c309421656f247af51a4b60e35c4b
